+++ indico/modules/events/editing/controllers/backend/editable_list.py
@@ -43,9 +43,12 @@ class RHEditableList(RHEditableTypeEditorBase):
 
     def _process_args(self):
         RHEditableTypeEditorBase._process_args(self)
+        revisions_strategy = joinedload('editables').selectinload('revisions')
+        revisions_strategy.selectinload('tags')
+        revisions_strategy.undefer('last_update_dt')
         self.contributions = (Contribution.query
                               .with_parent(self.event)
-                              .options(joinedload('editables').selectinload('revisions').selectinload('tags'),
+                              .options(revisions_strategy,
                                        joinedload('person_links'),
                                        joinedload('session'))
                               .order_by(Contribution.friendly_id)